Split TimePlugin out of CorePlugin
