+++ crates/bevy_core/Cargo.toml
@@ -16,7 +16,6 @@ bevy_ecs = { path = "../bevy_ecs", version = "0.8.0-dev", features = ["bevy_refl
 bevy_math = { path = "../bevy_math", version = "0.8.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.8.0-dev", features = ["bevy"] }
 bevy_tasks = { path = "../bevy_tasks", version = "0.8.0-dev" }
-bevy_time = { path = "../bevy_time", version = "0.8.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.8.0-dev" }
 
 # other
