+++ crates/bevy_core/src/lib.rs
@@ -15,26 +15,13 @@ pub mod prelude {
 }
 
 use bevy_app::prelude::*;
-use bevy_ecs::{
-    entity::Entity,
-    schedule::{ExclusiveSystemDescriptorCoercion, SystemLabel},
-    system::{IntoExclusiveSystem, ResMut},
-};
-use bevy_time::{FixedTimesteps, Time, Timer};
+use bevy_ecs::entity::Entity;
 use bevy_utils::HashSet;
 use std::ops::Range;
 
 /// Adds core functionality to Apps.
 #[derive(Default)]
 pub struct CorePlugin;
-
-/// A `SystemLabel` enum for ordering systems relative to core Bevy systems.
-#[derive(Debug, PartialEq, Eq, Clone, Hash, SystemLabel)]
-pub enum CoreSystem {
-    /// Updates the elapsed time. Any system that interacts with [Time] component should run after
-    /// this.
-    Time,
-}
 
 impl Plugin for CorePlugin {
     fn build(&self, app: &mut App) {
@@ -45,28 +32,15 @@ impl Plugin for CorePlugin {
             .unwrap_or_default()
             .create_default_pools(&mut app.world);
 
-        app.init_resource::<Time>()
-            .init_resource::<FixedTimesteps>()
-            .register_type::<HashSet<String>>()
+        app.register_type::<HashSet<String>>()
             .register_type::<Option<String>>()
             .register_type::<Entity>()
             .register_type::<Name>()
-            .register_type::<Range<f32>>()
-            .register_type::<Timer>()
-            // time system is added as an "exclusive system" to ensure it runs before other systems
-            // in CoreStage::First
-            .add_system_to_stage(
-                CoreStage::First,
-                time_system.exclusive_system().label(CoreSystem::Time),
-            );
+            .register_type::<Range<f32>>();
 
         register_rust_types(app);
         register_math_types(app);
     }
-}
-
-fn time_system(mut time: ResMut<Time>) {
-    time.update();
 }
 
 fn register_rust_types(app: &mut App) {
