+++ crates/bevy_internal/src/default_plugins.rs
@@ -3,6 +3,7 @@ use bevy_app::{PluginGroup, PluginGroupBuilder};
 /// This plugin group will add all the default plugins:
 /// * [`LogPlugin`](bevy_log::LogPlugin)
 /// * [`CorePlugin`](bevy_core::CorePlugin)
+/// * [`TimePlugin`](bevy_time::TimePlugin)
 /// * [`TransformPlugin`](bevy_transform::TransformPlugin)
 /// * [`HierarchyPlugin`](bevy_hierarchy::HierarchyPlugin)
 /// * [`DiagnosticsPlugin`](bevy_diagnostic::DiagnosticsPlugin)
@@ -27,6 +28,7 @@ impl PluginGroup for DefaultPlugins {
     fn build(&mut self, group: &mut PluginGroupBuilder) {
         group.add(bevy_log::LogPlugin::default());
         group.add(bevy_core::CorePlugin::default());
+        group.add(bevy_time::TimePlugin::default());
         group.add(bevy_transform::TransformPlugin::default());
         group.add(bevy_hierarchy::HierarchyPlugin::default());
         group.add(bevy_diagnostic::DiagnosticsPlugin::default());
@@ -76,6 +78,7 @@ impl PluginGroup for DefaultPlugins {
 
 /// Minimal plugin group that will add the following plugins:
 /// * [`CorePlugin`](bevy_core::CorePlugin)
+/// * [`TimePlugin`](bevy_time::TimePlugin)
 /// * [`ScheduleRunnerPlugin`](bevy_app::ScheduleRunnerPlugin)
 ///
 /// See also [`DefaultPlugins`] for a more complete set of plugins
@@ -84,6 +87,7 @@ pub struct MinimalPlugins;
 impl PluginGroup for MinimalPlugins {
     fn build(&mut self, group: &mut PluginGroupBuilder) {
         group.add(bevy_core::CorePlugin::default());
+        group.add(bevy_time::TimePlugin::default());
         group.add(bevy_app::ScheduleRunnerPlugin::default());
     }
 }
